+++ comics/comics/bugcomic.py
@@ -16,9 +16,8 @@ class Crawler(CrawlerBase):
     def crawl(self, pub_date):
         feed = self.parse_feed('http://www.bugcomic.com/feed/')
         for entry in feed.for_date(pub_date):
-            url = entry.summary.src('img[src*="/comics-rss/"]')
+            url = entry.summary.src('img[class="comicthumbnail"]')
+            title = entry.title
             if url is None:
                 return
-            url = url.replace('-rss', '')
-            title = entry.title
             return CrawlerImage(url, title)