+++ comics/comics/boxerhockey.py
@@ -9,16 +9,14 @@ class Meta(MetaBase):
     rights = 'Tyson "Rittz" Hesse'
 
 class Crawler(CrawlerBase):
-    time_zone = -7
     history_capable_days = 30
+    time_zone = -7
 
     def crawl(self, pub_date):
-        feed = self.parse_feed('http://boxerhockey.fireball20xl.com/inc/feed.php')
+        feed = self.parse_feed(
+            'http://boxerhockey.fireball20xl.com/inc/feed.php')
         for entry in feed.for_date(pub_date):
             title = entry.title
-            page_uri = entry.link
-            page = self.parse_page(page_uri)
+            page = self.parse_page(entry.link)
             url = page.src('img#comicimg')
-            text = None
-
-            return CrawlerImage(url, title, text)
+            return CrawlerImage(url, title)