+++ docs/changes.rst
@@ -23,6 +23,8 @@ In development
   site changes.
 - Update ``extralife`` to work after source site changes.
 - Update ``gunshow`` to work after source site changes.
+- Update ``questionablecontent`` to use site instead of feed, since it lacks
+  some releases.
 
 
 1.0.3 (2010-07-26)