+++ docs/changes.rst
@@ -19,6 +19,7 @@ In development
 - New: ``marriedtothesea``
 - New: ``stuffnoonetoldme``
 - New: ``utensokker``
+- Update ``boxerhockey`` to use feed instead of site.
 - Update ``bugcomic`` to not fail if URL is not empty, and to work after source
   site changes.
 - Update ``extralife`` to work after source site changes.